LayoutTests/fast/dom/dataset-name-getter-symbols-expected.txt

Test the name getters on DOMStringMap (dataset).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.body.dataset.webkitRocks === 'yes' is true
PASS Object.keys(document.body.dataset).length === 1 is true
PASS document.body.dataset[''] = 'test' did not throw exception.
PASS document.body.dataset[''] === 'test' is true
PASS document.body.dataset[Symbol()] === undefined is true
PASS Object.keys(document.body.dataset).length === 2 is true
PASS document.body.dataset[knownSymbol] === undefined is true
PASS Object.keys(document.body.dataset).length === 2 is true
PASS document.body.dataset[knownSymbol] = 'symbol property' did not throw exception.
PASS document.body.dataset[knownSymbol] === 'symbol property' is true
PASS delete document.body.dataset[knownSymbol] is true
PASS document.body.dataset[knownSymbol] === undefined is true
PASS Object.keys(document.body.dataset).length === 2 is true
PASS successfullyParsed is true

TEST COMPLETE
